+++ core/src/main/scala/ai/lum/odinson/MentionFactory.scala
@@ -1,7 +1,7 @@
 package ai.lum.odinson
 
 import ai.lum.odinson.lucene.OdinResults
-import ai.lum.odinson.lucene.search.OdinsonScoreDoc
+import ai.lum.odinson.lucene.search.{OdinsonIndexSearcher, OdinsonScoreDoc}
 import ai.lum.odinson.utils.MostRecentlyUsed
 import com.typesafe.config.Config
 
@@ -11,6 +11,26 @@ trait IdGetter {
   def getDocId: String
   def getSentId: String
 }
+
+class KnownIdGetter(docId: String, sentId: String) extends IdGetter {
+  override def getDocId: String = docId
+  override def getSentId: String = sentId
+}
+
+class LazyIdGetter(indexSearcher: OdinsonIndexSearcher, documentId: Int) extends IdGetter {
+  protected lazy val document = indexSearcher.doc(documentId)
+  protected lazy val docId: String = document.getField("docId").stringValue
+  protected lazy val sentId: String = document.getField("sentId").stringValue
+
+  def getDocId: String = docId
+
+  def getSentId: String = sentId
+}
+
+object LazyIdGetter {
+  def apply(indexSearcher: OdinsonIndexSearcher, docId: Int): LazyIdGetter = new LazyIdGetter(indexSearcher, docId)
+}
+
 
 trait MentionFactory {
